app/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
jinja2==3.1.5
psycopg2-binary==2.9.10
requests==2.31.0
PyYAML==6.0.1

